Move from Godeps to go modules
This commit removes the configuration files for Godeps as well as the
vendored dependencies, replacing them with go modules, Go's built-in
dependency management system. This dramatically slims down the size of
the repo (from 25M to 324K, discounting the .git directory) and greatly
speeds up cloning times. This will also provide mechanisms for managing
versions of any auxiliary tools (e.g. linters), creating a reproducible
environment for developers and CI/CD efforts.

This also modifies the Makefile to take into account that the repo no
longer needs to be cloned into the GOPATH.

Change-Id: I2213792cc3ce81831d5b835f2252ca6f137e0086